requirements.txt
web3
python-dotenv
py-etherscan-api
pandas
openai
langchain
simpleaichat
streamlit